adds ImagePreviewer component and useImageSrc hook
