+++ src/content/projects/you-wouldn-t-download-a-hedgehog.mdoc
@@ -25,9 +25,6 @@ Hedgehogs may be small, but they're full of personality! Unlike some other pets,
 {% ImagePopout
    image="/you-wouldn-t-download-a-hedgehog/filename with a space.jpg" /%}
 
-{% ImagePopout
-   image="/you-wouldn-t-download-a-hedgehog/filename with a space.jpg" /%}
-
 ### **Responsible Pet Ownership**
 
 While hedgehogs make fantastic pets, it's essential to remember that they require specialized care. From providing the right diet to creating a suitable habitat, there are several factors to consider before bringing home a hedgehog. Additionally, it's crucial to source your hedgehog responsibly from reputable breeders or adoption centers to ensure the health and well-being of these wonderful animals. Remember, you wouldn't download a hedgehog – so take the time to educate yourself on proper care practices.
